src/requirements.txt
Django==3.2.8
djangorestframework
psycopg2
uwsgi>=2.0.18, <3.0
gunicorn
django-cors-headers
bcrypt
django[argon2]
djangorestframework_simplejwt
djangorestframework-jwt
pyjwt==1.7.1
django-sslserver
opencv-python
opencv-contrib-python
opencv